+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Changelog
 
-### 1.5.0
+### 1.5.0 (30-May-2026)
 
 - **Added:**
-     - New monitor command `inspect`. Argument for `inspect` is a regular expression that selects  I/O registers. Then information and values are printed.
+     - New monitor command `ioregister`. If one additional argument is given, it will be matched against the I/O register names (using wildcard matching), and the matched registers and their values will be printed. If a unique register is addressed, then the bitfields (if any) are printed as well. If the expression matches a unique bitfield in a register, this will be printed. If two arguments are given, then the second should be an integer value, which is stored in the unique register or bitfield referred to by the first argument.
      - The addsvd.py utility has been added in `deviceinfo`. This utility reads an SVD file and adds it as a Python data structure to the corresponding device file in `deviceinfo.devices`.  This has to be called after harvest.py.
-
 - **Changed:**
      -  `sram_masked_read` and `sram_masked_write` are moved to xavrdebug, so that they can be used in the monitor command.
+- **Fixed:**
+     - Read/write masking did not work because the ronly/wonly lists were set before the `device_info` was read.
+
 
 ### 1.4.0 (23-May-2026)